requirements.txt
aiosqlite==0.17.0
anyio==3.6.1
bcrypt==4.0.0
certifi==2022.9.24
charset-normalizer==2.1.1
fastapi==0.85.0
finnhub-python==2.4.14
h11==0.14.0
h2==4.1.0
hpack==4.0.0
hypercorn==0.14.3
hyperframe==6.0.1
idna==3.4
iso8601==1.1.0
passlib==1.7.4
priority==2.0.0
pydantic==1.10.2
PyJWT==2.5.0
pypika-tortoise==0.1.6
python-multipart==0.0.5
pytz==2022.4
requests==2.28.1
six==1.16.0
sniffio==1.3.0
starlette==0.20.4
toml==0.10.2
tortoise-orm==0.19.2
typing_extensions==4.3.0
urllib3==1.26.12
wsproto==1.2.0
